feat(issue-8): Implement LawProposalCreateSerializer for proposal creation
- Added LawProposalCreateSerializer to handle request body for POST /laws/proposals/.
- Included validation to prevent using LAW_ID_CONST for law proposals.
- Updated views to utilize the new serializer for proposal creation.
